Function extension final and basic testing
Extends Function to correctly return the input number as roman numbers. Now supports the rule that no roman letter may occur 4 times after each other, with the help of an extended lookup table.
Also there is now basic testing for the input. Only numbers in the range of 1-3999 may be processed.